+++ README.md
@@ -28,7 +28,7 @@ $ flutter packages get
 3. Import it: Now in your Dart code, you can use:
 
 ```dart
-import 'package:sqlbrite/sqlbrite.dart';
+import 'package:sqlbrite/type_defs.dart';
 ```
 
 Usage
