+++ analysis_options.yaml
@@ -1,0 +1,5 @@
+include: package:pedantic/analysis_options.1.8.0.yaml
+linter:
+  rules:
+    - public_member_api_docs
+    - prefer_final_locals
