+++ lib/sqlbrite.dart
@@ -1,8 +1,7 @@
-library sqlbrite;
-
 ///
 /// Author: Petrus Nguyễn Thái Học
 /// Github: https://github.com/hoc081098
 ///
+library sqlbrite;
 
-export 'src/sqlbrite.dart';
+export 'src/type_defs.dart';
