+++ lib/src/stream_transformers/query_to_list_stream_transformer.dart
@@ -1,0 +1,78 @@
+import 'dart:async';
+
+import '../../sqlbrite.dart';
+
+///
+class QueryToListStreamTransformer<T>
+    extends StreamTransformerBase<Query, List<T>> {
+  final StreamTransformer<Query, List<T>> _transformer;
+
+  ///
+  QueryToListStreamTransformer(T mapper(Map<String, dynamic> row))
+      : _transformer = _buildTransformer(mapper);
+
+  @override
+  Stream<List<T>> bind(Stream<Query> stream) => _transformer.bind(stream);
+
+  static StreamTransformer<Query, List<T>> _buildTransformer<T>(
+      T mapper(Map<String, dynamic> row)) {
+    return StreamTransformer<Query, List<T>>((
+      Stream<Query> input,
+      bool cancelOnError,
+    ) {
+      StreamController<List<T>> controller;
+      StreamSubscription<Query> subscription;
+
+      add(List<Map<String, dynamic>> rows) {
+        final items = rows.map((row) => mapper(row)).toList(growable: false);
+        controller.add(items);
+      }
+
+      onListen() {
+        subscription = input.listen(
+          (Query event) {
+            Future<List<Map<String, dynamic>>> newValue;
+            try {
+              newValue = event();
+            } catch (e, s) {
+              controller.addError(e, s);
+              return;
+            }
+            if (newValue != null) {
+              subscription.pause();
+              newValue
+                  .then(add, onError: controller.addError)
+                  .whenComplete(subscription.resume);
+            }
+          },
+          onError: controller.addError,
+          onDone: controller.close,
+        );
+      }
+
+      if (input.isBroadcast) {
+        controller = StreamController<List<T>>.broadcast(
+          onListen: onListen,
+          onCancel: () {
+            subscription.cancel();
+          },
+          sync: true,
+        );
+      } else {
+        controller = StreamController<List<T>>(
+          onListen: onListen,
+          onPause: () {
+            subscription.pause();
+          },
+          onResume: () {
+            subscription.resume();
+          },
+          onCancel: () => subscription.cancel(),
+          sync: true,
+        );
+      }
+
+      return controller.stream.listen(null);
+    });
+  }
+}
