+++ lib/src/type_defs.dart
@@ -1,0 +1,7 @@
+import 'dart:async';
+
+///
+typedef Future<List<Map<String, dynamic>>> Query();
+
+///
+typedef void Logger(String message);
